move file to correct location
