+++ src/briefcase/console.py
@@ -960,3 +960,45 @@ class Console:
         )
 
         return ordered[int(index) - 1][0]
+
+    def boolean_question(
+        self,
+        description,
+        intro: str,
+        default: bool | None = None,
+        override_value: bool | None = None,
+    ) -> bool:
+        """Ask the user a boolean question who's answer requires selecting yes/no
+
+        :param description: A short description of the question being asked. This text
+            is used in prompts and a header bar prefacing the question.
+        :param intro: An introductory paragraph explaining the question being asked.
+        :param default: The default option for empty user input.
+        :param override_value: A pre-selected answer for the question. This can be used
+            to shortcut asking the question, such as when a command line option provides
+            a value. If provided and valid, the header bar will be displayed, but the
+            intro paragraph and option list will not.
+        :returns: The user's chosen answer or none if closed without input
+        """
+        if default not in {True, False, None}:
+            raise ValueError(
+                f"'{default}' is not a valid default value for {description}? Must be True, False, or None."
+            )
+
+        self.divider(title=description)
+
+        if override_value is not None:
+            self.print()
+            self.print(f"Using override value {override_value!r}")
+            if isinstance(override_value, bool):
+                return override_value
+            else:
+                raise ValueError(
+                    f"Invalid override value for {description}: must be True or False"
+                )
+
+        self.prompt()
+        self.prompt(self.textwrap(intro))
+        self.prompt()
+
+        return self.input_boolean(description, default=default)
